FPLA-290 move docmosis related classes to appropriate package
