+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/docmosis/DocmosisFactorsParenting.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.model;
+package uk.gov.hmcts.reform.fpl.model.docmosis;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
