+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/casesubmission/CaseSubmissionTemplateDataGenerationServiceTest.java
@@ -22,11 +22,6 @@ import uk.gov.hmcts.reform.fpl.model.ApplicantParty;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
 import uk.gov.hmcts.reform.fpl.model.Child;
 import uk.gov.hmcts.reform.fpl.model.ChildParty;
-import uk.gov.hmcts.reform.fpl.model.DocmosisFactorsParenting;
-import uk.gov.hmcts.reform.fpl.model.DocmosisHearing;
-import uk.gov.hmcts.reform.fpl.model.DocmosisHearingPreferences;
-import uk.gov.hmcts.reform.fpl.model.DocmosisInternationalElement;
-import uk.gov.hmcts.reform.fpl.model.DocmosisRisks;
 import uk.gov.hmcts.reform.fpl.model.Grounds;
 import uk.gov.hmcts.reform.fpl.model.GroundsForEPO;
 import uk.gov.hmcts.reform.fpl.model.Orders;
@@ -42,6 +37,11 @@ import uk.gov.hmcts.reform.fpl.model.common.EmailAddress;
 import uk.gov.hmcts.reform.fpl.model.common.Telephone;
 import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisApplicant;
 import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisCaseSubmission;
+import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisFactorsParenting;
+import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisHearing;
+import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisHearingPreferences;
+import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisInternationalElement;
+import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisRisks;
 import uk.gov.hmcts.reform.fpl.service.UserDetailsService;
 import uk.gov.hmcts.reform.fpl.service.config.LookupTestConfig;
 import uk.gov.hmcts.reform.fpl.utils.FixedTimeConfiguration;
